Update of table "dataframe"
This update changes how many significant numbers are shown in the report table since the measure is in CLP zero are displayed

+++ writeReport.py
@@ -73,7 +73,8 @@ def writeReport(client,cryptoCurrency,startTime,records):
       table.dataframe tbody tr th {text-align: center;}
     </style>
     """
-
+    records["Total"]=records["Total"].astype(int)
+    records["Fees"]=records["Fees"].astype(int)
     html+=records.to_html().replace('<tr style="text-align: right;">','<tr>')
 
     # Write the report in PDF
